Refactor player actions, reducers, and state.
