+++ src/garden/gardenContainer.js
@@ -1,13 +1,16 @@
 import { connect } from 'react-redux';
+
 import GardenComponent from './gardenComponent';
 import {
-  agePlant,
   increaseWeedQuantity,
+} from '../player/playerActions';
+import {
+  agePlant,
   removePlant,
   addNotification
 } from '../state/actions';
-import { getStrainById } from '../utils/weedUtils';
 import { BASE_HARVEST_PER_PLANT } from '../utils/constants';
+import { getStrainById } from '../utils/weedUtils';
 
 const mapStateToProps = state => {
   return {
