+++ src/player/player.css
@@ -1,4 +1,4 @@
-.playerStatus {
+.player {
   flex-basis: 50%;
   padding-right: 10px;
 }
